+++ src/__init__.lua
@@ -15,3 +15,5 @@ import "TravelWindowII.src.OptionsWindow";
 import "TravelWindowII.src.TravelButton";
 -- import "TravelWindowII.src.TravelTests";
 import "TravelWindowII.src.OptionsPanel";
+import "TravelWindowII.src.ChangelogData";
+import "TravelWindowII.src.UpdateNotificationWindow";