+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## v4.0.0
+
+- "MAJOR: introducing a brand new view mode called Map Mode",
+- "MAJOR: check it out by right-clicking the TW II window, select MODE and then MAP VIEW",
+- "enhance: adding some region labels to U46 skill data for EN/DE/FR/RU",
+- "enhance: vendor and skillname updates from U46.0.1 for FR"
+
 ## v3.2.0
 
 - enhance: window positioning is happening now relative to screen rather than absolute - positioning will work now much better if screen resolutions changes